+++ R/image_reg.R
@@ -290,8 +290,16 @@ get_voi_seg <- function(voi, mri_seg) {
 get_voi_seg_psf <- function(psf, mri_seg) {
   # check the image orientation etc is the same
   check_geom(psf, mri_seg)
+  
+  # check the segmentation data has only integer values
+  if (!all(mri_seg == as.integer(mri_seg))) {
+    stop("Segmentation file has non-integer values.") 
+  }
+  
   mask <- (abs(psf) > 0)
   vals <- mri_seg[mask]
+  
+  
   other <- sum(as.numeric(vals == 0) * psf[mask])
   csf   <- sum(as.numeric(vals == 1) * psf[mask])
   gm    <- sum(as.numeric(vals == 2) * psf[mask])
@@ -469,6 +477,11 @@ get_mrsi2d_seg <- function(mrs_data, mri_seg, ker) {
   # reformat seg data into the same space as the full MRSI voi
   voi <- get_mrsi_voi(mrs_data)
   mri_seg_crop <- resample_img(mri_seg, voi, interp = 0L)
+  
+  # check mri_seg_crop has only integer values
+  if (!all(mri_seg_crop == as.integer(mri_seg_crop))) {
+    stop("Segmentation file has non-integer values.") 
+  }
   
   mri_seg_0 <- rowMeans(mri_seg_crop == 0, dims = 2)
   mri_seg_1 <- rowMeans(mri_seg_crop == 1, dims = 2)